Additional plumbing to connect the OAuth login to the DataStore.
Move InMemory data store support/backend to a different file.
